refactor: add detect point component
