+++ src/views/components/topology/topo-aside.vue
@@ -22,8 +22,8 @@
       <use xlink:href="#issues"/>
     </svg>
     <svg v-if="showServerInfo" class="link-topo-aside-btn icon cp lg"
-         @click="show = !show"
-         :style="`position:absolute;left:290px;transform: rotate(${show?0 : 180}deg);top:50px;`">
+      @click="show = !show"
+      :style="`position:absolute;left:290px;transform: rotate(${show?0 : 180}deg);top:50px;`">
       <use xlink:href="#chevron-left"/>
     </svg>
     <TopoService/>
@@ -42,118 +42,23 @@
         </div>
         <div>
           <TopoChart
-              v-if="rocketDashboard.serviceApdexScore.ApdexScore.length"
-              :data="rocketDashboard.serviceApdexScore.ApdexScore"
-              :intervalTime="intervalTime"
-              title="Service ApdexScore"
-              unit=""
+            v-if="rocketDashboard.serviceApdexScore.ApdexScore.length"
+            :data="rocketDashboard.serviceApdexScore.ApdexScore"
+            :intervalTime="intervalTime"
+            title="Service ApdexScore"
+            unit=""
           />
           <TopoChart
-              v-if="rocketDashboard.serviceSLA.SLA.length"
-              :data="rocketDashboard.serviceSLA.SLA"
-              :intervalTime="intervalTime"
-              :precent="true"
-              title="Service SLA"
-              unit="%"
+            v-if="rocketDashboard.serviceSLA.SLA.length"
+            :data="rocketDashboard.serviceSLA.SLA"
+            :intervalTime="intervalTime"
+            :precent="true"
+            title="Service SLA"
+            unit="%"
           />
         </div>
       </div>
-      <div v-if="stateTopo.selectedCallId || showServerInfo"
-           :class="`link-topo-aside-box link-topo-aside-box-${isMini?'min':'max'}`"
-           :style="`top:80px;position: fixed;right: 30px;${showInfoCount === 0 ? 'animation: unset;': ''}`">
-        <svg
-            :style="`position:absolute;left:-48px;top:0;transform: rotate(${isMini?0 : 180}deg);`"
-            class="link-topo-aside-btn icon cp lg"
-            @click="setShowInfo"
-        >
-          <use xlink:href="#chevron-left"/>
-        </svg>
-        <div class="mb-5 clear">
-          <span v-if="stateTopo.selectedCallId" class="b dib mr-20 vm">{{ $t('detectPoint') }}</span>
-          <span v-else-if="showServerInfo" class="b dib mr-20 vm">{{ $t('serviceDetail') }}</span>
-          <span
-              v-if="stateTopo.detectPoints.indexOf('CLIENT') !== -1"
-              :class="{'active':!stateTopo.mode}"
-              class="link-topo-aside-box-btn tc r sm cp b"
-              @click="setMode(false)"
-          >{{ this.$t('client') }}</span>
-          <span
-              v-if="stateTopo.detectPoints.indexOf('SERVER') !== -1"
-              :class="{'active':stateTopo.mode}"
-              class="link-topo-aside-box-btn tc r sm cp b"
-              @click="setMode(true)"
-          >{{ this.$t('server') }}</span>
-        </div>
-        <div v-if="showInfo">
-          <div v-if="stateTopo.selectedCallId">
-            <TopoChart
-                v-if="stateTopo.getResponseTimeTrend.length"
-                :data="stateTopo.getResponseTimeTrend"
-                :intervalTime="intervalTime"
-                :title="$t('avgResponseTime')"
-                unit="ms"
-            />
-            <TopoChart
-                v-if="stateTopo.getThroughputTrend.length"
-                :data="stateTopo.getThroughputTrend"
-                :intervalTime="intervalTime"
-                :title="$t('avgThroughput')"
-                unit="cpm"
-            />
-            <TopoChart
-                v-if="stateTopo.getSLATrend.length"
-                :data="stateTopo.getSLATrend"
-                :intervalTime="intervalTime"
-                :precent="true"
-                :title="$t('avgSLA')"
-                unit="%"
-            />
-            <ChartResponse
-                v-if="stateTopo.p50.length"
-                :data="stateTopo"
-                :intervalTime="intervalTime"
-                :title="$t('percentResponse')"
-            />
-          </div>
-          <div v-else-if="showServerInfo">
-            <TopoChart
-                v-if="rocketDashboard.serviceResponseTime.ResponseTime.length"
-                :data="rocketDashboard.serviceResponseTime.ResponseTime"
-                :intervalTime="intervalTime"
-                title="Service ResponseTime"
-                unit="ms"
-            />
-            <TopoChart
-                v-if="rocketDashboard.serviceThroughput.Throughput.length"
-                :data="rocketDashboard.serviceThroughput.Throughput"
-                :intervalTime="intervalTime"
-                title="Service Throughput"
-                unit="cpm"
-            />
-            <ChartResponse
-                v-if="rocketDashboard.servicePercent.p50.length"
-                :data="rocketDashboard.servicePercent"
-                :intervalTime="intervalTime"
-                title="Service Response Time Percentile"
-                unit="ms"
-            />
-          </div>
-        </div>
-        <div class="show-dependency" v-if="stateTopo.selectedCallId">
-          <a class="rk-btn lg" @click="dialogTopoVisible=true">{{$t('ShowInstanceDependency')}}</a>
-          <el-dialog
-            class="instance-dependency" 
-            :width="'90%'"
-            :title="$t('InstanceDependencyTitle')"
-            :visible.sync="dialogTopoVisible"
-            :modal-append-to-body="false"
-            :close-on-click-modal="false"
-            destroy-on-close
-          >
-            <TopoInstanceDependency :selectedCallId="stateTopo.selectedCallId" />
-          </el-dialog>
-        </div>
-      </div>
+      <TopoDetectPoint :rocketDashboard="rocketDashboard" :instanceBtnIsShow="true"/>
     </div>
     <el-drawer
         v-if="stateTopo.showAlarmDialog"
@@ -164,9 +69,9 @@
         :modal-append-to-body="false"
     >
       <alarm-containers :style="`height: ${drawerMainBodyHeight}`"
-                        :alarmScope="{label: 'Service', key: 'Service'}"
-                        inTopo
-                        :keyword="stateTopo.honeycombNode.name"
+        :alarmScope="{label: 'Service', key: 'Service'}"
+        inTopo
+        :keyword="stateTopo.honeycombNode.name"
       />
     </el-drawer>
     <el-drawer
@@ -178,19 +83,19 @@
         :modal-append-to-body="false"
     >
       <trace-containers :style="`height: ${drawerMainBodyHeight}`"
-                        :service="{label: stateTopo.honeycombNode.name, key: stateTopo.honeycombNode.id}"
-                        inTopo
+        :service="{label: stateTopo.honeycombNode.name, key: stateTopo.honeycombNode.id}"
+        inTopo
       />
     </el-drawer>
     <instances-survey-window
-        v-if="stateTopo.showInstancesDialog"
-        :is-show.sync="stateTopo.showInstancesDialog"
-        :instances="stateDashboardOption.instances"
+      v-if="stateTopo.showInstancesDialog"
+      :is-show.sync="stateTopo.showInstancesDialog"
+      :instances="stateDashboardOption.instances"
     />
     <endpoint-survey-window
-        v-if="stateTopo.showEndpointDialog"
-        :is-show.sync="stateTopo.showEndpointDialog"
-        :endpoints="stateDashboardOption.endpoints"
+      v-if="stateTopo.showEndpointDialog"
+      :is-show.sync="stateTopo.showEndpointDialog"
+      :endpoints="stateDashboardOption.endpoints"
     />
   </aside>
 </template>
@@ -208,12 +113,13 @@
   import ChartResponse from './topo-response.vue';
   import TopoService from './topo-services.vue';
   import TopoInstanceDependency from './topo-instance-dependency.vue';
+  import TopoDetectPoint from './topo-detect-point.vue';
 
   @Component({
     components: {
       TopoChart, TopoService, ChartResponse, Radial, AlarmContainers,
       TraceContainers, InstancesSurveyWindow, EndpointSurveyWindow,
-      TopoInstanceDependency,
+      TopoInstanceDependency, TopoDetectPoint,
     },
   })
   export default class TopoAside extends Vue {
@@ -224,9 +130,6 @@
     @Action('rocketTopo/GET_TOPO') private GET_TOPO: any;
     @Action('rocketTopo/CLEAR_TOPO') private CLEAR_TOPO: any;
     @Action('SELECT_SERVICE') private SELECT_SERVICE: any;
-    @Action('GET_QUERY') private GET_QUERY: any;
-    @Action('MIXHANDLE_CHANGE_GROUP_WITH_CURRENT') private MIXHANDLE_CHANGE_GROUP_WITH_CURRENT: any;
-    @Action('MIXHANDLE_GET_OPTION') private MIXHANDLE_GET_OPTION: any;
     @Mutation('rocketTopo/SET_MODE_STATUS') private SET_MODE_STATUS: any;
     @Action('rocketTopo/CLEAR_TOPO_INFO') private CLEAR_TOPO_INFO: any;
     @Mutation('SET_COMPS_TREE') private SET_COMPS_TREE: any;
@@ -238,9 +141,6 @@
     private initState = initState;
     private radioStatus: boolean = false;
     private show: boolean = true;
-    private showInfo: boolean = false;
-    private isMini: boolean = true;
-    private showInfoCount: number = 0;
 
     private get showServerInfo() {
       return this.stateTopo.currentNode.name && this.stateTopo.currentNode.isReal;
@@ -288,54 +188,8 @@
       this.getTopo();
     }
 
-    @Watch('stateTopo.selectedCallId')
-    private watchDetectPointNodeId(newValue: string) {
-      if (newValue || this.stateTopo.currentNode.isReal) {
-        this.showInfo = true;
-      } else {
-        this.showInfo = false;
-        this.isMini = true;
-      }
-    }
-
-    @Watch('stateTopo.currentNode.name')
-    private watchCurrentNodeIsReal(newValue: boolean) {
-      const service = this.stateTopo.currentNode;
-      if (this.stateTopo.currentNode.isReal) {
-        this.MIXHANDLE_CHANGE_GROUP_WITH_CURRENT({index: 0, current: 1});
-        this.MIXHANDLE_GET_OPTION({compType: 'service', duration: this.durationTime})
-          .then(() => {
-            this.GET_QUERY({
-              serviceId: service.id || '',
-              duration: this.durationTime,
-            });
-          });
-      }
-      if (newValue || this.stateTopo.selectedCallId) {
-        this.showInfo = true;
-      } else {
-        this.showInfo = false;
-        this.isMini = true;
-      }
-    }
-
-
     private showRadial() {
       this.radioStatus = !this.radioStatus;
-    }
-
-    private setMode(mode: boolean) {
-      this.SET_MODE_STATUS(mode);
-      this.stateTopo.callback();
-    }
-
-    private setShowInfo() {
-      this.showInfo = false;
-      this.showInfoCount = 1;
-      this.isMini = !this.isMini;
-      setTimeout(() => {
-        this.showInfo = true;
-      }, 550);
     }
   }
 </script>
@@ -350,36 +204,6 @@
     .to-apm {
       padding-top: 10px;
       border-top: 1px solid #d8d8d866;
-    }
-  }
-
-  .link-topo-aside-box-btn {
-    color: #626977;
-    border: 1px solid;
-    padding: 0px 3px;
-    width: 45px;
-    display: inline-block;
-
-    &.active {
-      color: #448dfe;
-    }
-  }
-  .show-dependency {
-    margin: 20px 0;
-    .rk-btn {
-      display: block;
-      text-align: center;
-    }
-    .el-dialog__header{
-      background: #333;
-      .el-dialog__title{
-        color: #efeff1;
-      }
-    }
-    .el-dialog__body{
-      background: #333;
-      color: #efeff1;
-      height: 650px;
     }
   }
